edited descriptions
I accidentally edited the commented out code first (hopefully not a big issue?), but I also edited the actual code and changed the who we are paragraph, deleted the why section, and edited the progress section. I left a note in there to add a button in the progress section to link to the classes page.